requirements.txt
appdirs==1.4.3
boto==2.46.1
configparser==3.5.0
Django==1.11
django-braces==1.11.0
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2
enum34==1.1.6
flake8==3.3.0
mccabe==0.6.1
oauthlib==2.0.1
packaging==16.8
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
